requirements.txt
fastapi
uvicorn
pillow
scikit-learn
python-multipart

